fix: bug preventing new segments from being savable (#823)
* fix: bug preventing new segments from being saveable

* fix: rename function

* fix: console error when using history.block before segment re-render

Previously, if the conditions to show prompt are not met in usePrompt, we set the ref to null, which results in a JS console error if trying to save a dirty segment on the first component render. Now, we simply don't destroy the unblock function in that case, which fixes the issue. However, there's still a bug in the React Router History library which sets the current URL before the user has a chance to cancel history change from the history.block prompt.

Co-authored-by: Jefferson Rabb <jeff@atavist.net>