fix: prevent coauthors capability check infinite loop (#46)
This fixes an issue where, when NP Network, Distributor, CAP, and NP Newsletters are active, sending a newsletters causes an infinite capability check which results in a fatal memory error.

The issue happens because we are creating a Distributor Post object to check a distributed posts linked status when filtering coauthors via the get_coauthors hook. But this same hook is eventually triggered in the instantiation of the Distributor Post object.

This PR fixes this by checking a posts linked status directly via post meta.