fix: register with empty name fields (#4175)
This PR fixes an issue where we were incorrectly setting a readers first name to the default nicename generated from the readers email on registration. As a result, Woo would not update the first name during checkout since Woo skips this step in favor of any name data that is already set.